requirements_old.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asyncio-mqtt==0.16.2
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
Bottleneck==1.5.0
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorlog==6.9.0
cryptography==45.0.4
cssselect==1.3.0
cssutils==2.11.1
dashscope==1.23.6
dataclasses-json==0.6.7
decorator==5.2.1
distro==1.9.0
docstring_parser==0.16
easyocr==1.7.2
einops==0.8.1
et_xmlfile==2.0.0
faiss-cpu==1.11.0
fastapi==0.115.13
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
frozenlist==1.7.0
fsspec==2025.5.1
ftfy==6.3.1
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-aiplatform==1.98.0
google-cloud-bigquery==3.34.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.21.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
GPUtil==1.4.0
greenlet==3.2.3
grpc-google-iam-v1==0.14.2
grpcio==1.73.0
grpcio-status==1.73.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.0
idna==3.10
imageio==2.37.0
imagesize==1.4.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-community==0.3.26
langchain-core==0.3.66
langchain-deepseek==0.1.3
langchain-google-vertexai==2.0.26
langchain-huggingface==0.3.0
langchain-openai==0.3.25
langchain-text-splitters==0.3.8
langgraph==0.4.8
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
langsmith==0.4.1
lazy_loader==0.4
lxml==6.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.5.1
mypy_extensions==1.1.0
networkx==3.5
ninja==1.11.1.4
numexpr==2.11.0
numpy==2.2.6
openai==1.91.0
opencv-contrib-python==4.10.0.84
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
opt-einsum==3.3.0
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
paddlex==3.1.3
paho-mqtt==2.1.0
pandas==2.3.1
pillow==11.3.0
premailer==3.10.0
prettytable==3.16.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
py-cpuinfo==9.0.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyJWT==2.10.1
pypdfium2==4.30.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rsa==4.9.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.16.0
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
sqlmodel==0.0.24
starlette==0.46.2
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2025.6.11
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ujson==5.10.0
urllib3==2.5.0
uvicorn==0.34.3
validators==0.35.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0
